create apiVersions static file
We want to include an apiVersionsUrl in the response to a new
`GET /about` call on Metasys. It would link to this file. The idea being that we likely will forget to update this table before a new Metasys release, but we can update this public file at any time.